Update Record.pm
in the current implementation, the constructor merges adjacent sections when opening the original hex file. 
i have modified the write function so that it still merges the adjacent sections when it is called, unless this section was already existing in the original hex file (after a 02 or 04 code is found). 
what do you think about this idea? 
in our system we needed this sections to be present in the final hex file. when two of the sections were adjacent, our system was not working properly. we have struggled a little bit until we found out the reason why. 
thank you in advance!